infra/requirements.txt
aws-cdk-lib==2.59.0
constructs>=10.0.0,<11.0.0
python-dotenv==0.21.0
pytest==7.2.0